refactor: projectApply 내부 인증검사 제거
